+++ lib/src/async_notifier.dart
@@ -3,12 +3,6 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
-/// Signature for callbacks that report data.
-typedef DataChanged<T> = void Function(T data);
-
-/// Signature for callbacks that report errors.
-typedef ErrorCallback = void Function(Object error, StackTrace stackTrace);
-
 /// Signature for a read-only [AsyncSnapshot] listener.
 typedef AsyncListenable<T> = ValueListenable<AsyncSnapshot<T>>;
 
@@ -23,24 +17,17 @@ class AsyncNotifier<T> extends ValueNotifier<AsyncSnapshot<T>> {
   /// - `error`: The initial [AsyncSnapshot.error].
   /// - `stackTrace`: The initial [AsyncSnapshot.stackTrace].
   /// - `state`: The initial [AsyncSnapshot.connectionState].
-  /// - `onData`: Invoked when new data is available.
-  /// - `onError`: Invoked when an error occurs during an asynchronous operation.
+  /// - `cancelOnError`: Whether the subscription should be canceled when an error occurs.
   ///
   /// Example:
   /// ```dart
-  /// final notifier = AsyncNotifier(<Todo>[], onData: (todos) {
-  ///   print('Todos changed: ${todos.length}');
-  /// }, onError: (error, stackTrace) {
-  ///   print('An error occurred: $error');
-  /// });
+  /// final notifier = AsyncNotifier<List<Todo>>();
   /// ```
   AsyncNotifier({
     T? data,
     Object? error,
     StackTrace stackTrace = StackTrace.empty,
     ConnectionState state = ConnectionState.none,
-    this.onData,
-    this.onError,
     this.cancelOnError,
   })  : assert(data == null || error == null),
         super(
@@ -51,12 +38,6 @@ class AsyncNotifier<T> extends ValueNotifier<AsyncSnapshot<T>> {
                   : AsyncSnapshot<T>.nothing().inState(state),
         );
 
-  /// Called when [AsyncSnapshot] resolves with new data.
-  final DataChanged<T>? onData;
-
-  /// Called whenever [AsyncSnapshot] resolves with error.
-  final ErrorCallback? onError;
-
   /// Whether the subscription should be canceled when an error occurs.
   final bool? cancelOnError;
 
@@ -65,23 +46,8 @@ class AsyncNotifier<T> extends ValueNotifier<AsyncSnapshot<T>> {
   StreamSubscription<T>? _subscription;
 
   @override
-  @visibleForTesting
   @Deprecated('Use `snapshot` instead')
   AsyncSnapshot<T> get value => super.value;
-
-  @override
-  @visibleForTesting
-  set value(AsyncSnapshot<T> value) {
-    final previous = super.value;
-    final next = super.value = value;
-
-    if (next.data is T && next.data != previous.data && !next.hasError) {
-      onData?.call(next.data as T);
-    }
-    if (next.hasError && next.error != previous.error) {
-      onError?.call(next.error!, next.stackTrace!);
-    }
-  }
 
   /// The [Future] currently being listened to.
   Future<T>? get future => _future;
@@ -94,7 +60,7 @@ class AsyncNotifier<T> extends ValueNotifier<AsyncSnapshot<T>> {
   /// Stop listening to any existing future and listens to the new one.
   /// - On start: Updates [snapshot] to 'waiting'.
   /// - On completion: Updates [snapshot] with done state and data.
-  /// - On error: Updates [snapshot] with error and invokes [onError].
+  /// - On error: Updates [snapshot] with error.
   ///
   /// No action is taken if the new future is identical to the current one.
   ///
@@ -128,14 +94,14 @@ class AsyncNotifier<T> extends ValueNotifier<AsyncSnapshot<T>> {
   /// Unsubscribes from any existing stream and subscribes to the new one.
   /// - On start: Updates [snapshot] to 'waiting'.
   /// - On new data: Updates [snapshot] with active state and data.
-  /// - On error: Updates [snapshot] with error and invokes [onError].
+  /// - On error: Updates [snapshot] with error.
   /// - On done: Updates [snapshot] to done state.
   ///
   /// No action is taken if the new stream is identical to the current one.
   ///
   /// Example:
   /// ```dart
-  /// final _user = AsyncNotifier.late<User>();
+  /// final _user = AsyncNotifier<User>();
   /// _user.stream = someUserStream;
   /// ```
   set stream(Stream<T>? stream) {